Initial Backend and Frontend, Register and Login with roles

+++ backend/controllers/userController.js
@@ -21,11 +21,14 @@ export const getUsers = async(req, res) => {
 
 export const deleteUser = async (req, res) => {
     try {
-        const user = await User.findByIdAndDelete(req.params.id)
-
+        const user = await User.findById(req.params.id)
         if(!user) {
             return res.status(404).json({message: "User not found"})
         }
+        if(user.role === "admin") {
+            return res.status(403).json({message: "Cannot delete admin users"})
+        }
+        await User.findByIdAndDelete(req.params.id)
         return res.status(200).json({message: "User deleted successfully"})
     } catch (error) {
         console.error("Error deleting user:", error)
